feat(ci): configured ci with C++ and Commit Linting
added a clang-tidy file containing a set of checks from google llvm and the core guideline, further down the line might add high performance checks, however currently they should no be needed.
The clang-format file is the google style guide for C++ as that seems to be a good style, might consider changing it down the line if a good reason comes up.
Added CircleCI configuration Yaml file for run the CI for the project.
Then configured CI to use clang-tidy to run static analysis of the code to check for code issues and possible bugs.
Also configured CI to use clang-format as the linter of the code.
Finally configured CI to use commitlint to lint commit messages.